Rare: Move sub-commands implementations in the `commands` subfolder
Right now we have two commands, and with possible third one comming soon.
