+++ rare/commands/launcher/console_dialog.py
