+++ rare/commands/launcher/lgd_helper.py
